refactor(vscode-extension): share offset->line/column helpers, fix provider-chain loc
